requirements.txt
configargparse>=1.4.1
hail
pygraphviz
google-cloud-storage

mock
tox
tox-venv   # work around tox virtualenv issues
sphinxcontrib-napoleon
